requirements.txt
RPi.GPIO==0.6.2
RPLCD==0.3.0
